Added weapon swapping functionality to change between pistol, machine gun and shotgun all with their own variables.

+++ Pengun/Assets/GromEnergyBarController.cs
@@ -27,7 +27,7 @@ public class GromEnergyBarController : MonoBehaviour
         UpdateGromEnergy();
     }
 
-    public void DecreaseGromEnergy(int amount)
+    public void DecreaseGromEnergy(float amount)
     {
         currentGromEnergy -= amount;
